requirements.txt
flask
feedparser
requests
mysql-connector-python